requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.3
Pillow==5.4.1
pytz==2018.9
django-leaflet==0.26.0